requirements.txt
flask>=0.10
lxml>=2.2
requests
requests_oauthlib
requests_toolbelt
pycrypto
